Refactored fetchUri
